Move logic into a hook.
